Change cast to fix warning: assuming signed overflow does not occur when assuming that (X - c) <= X is always true

+++ src/unc_text.h
@@ -137,7 +137,7 @@ public:
 
    int operator[](int idx) const
    {
-      return(((idx >= 0) && (idx < (int)m_chars.size())) ? m_chars[idx] : 0);
+      return(((idx >= 0) && ((size_t)idx < m_chars.size())) ? m_chars[idx] : 0);
    }
 
    /* throws an exception if out of bounds */
